pyproject.toml
[project]
name = "django-anchor"
dynamic = ["version"]
description = "Attach files to Django models"
readme = "README.md"
authors = [
    { name = "Elias Hernandis", email = "elias@hernandis.me" }
]
license = { name = "MIT", file = "LICENSE.md" }
requires-python = ">=3.10"
keywords = ["django", "files", "attachments"]
classifiers = [
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Topic :: File Formats",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "django>=4.2,<6",
    "pillow>=9.5,<12",
]

[project.urls]
Homepage = "https://github.com/knifecake/django-anchor"
Source = "https://github.com/knifecake/django-anchor"
Documentation = "https://django-anchor.readthedocs.io/en/latest/"
Tracker = "https://github.com/knifecake/django-anchor/issues"
Releases = "https://github.com/knifecake/django-anchor/releases"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "coverage>=7.6.9",
    "django-environ>=0.11.2",
    "django-storages[s3]>=1.14.4",
    "freezegun>=1.5.1",
    "pre-commit>=4.0.1",
    "requests>=2.32.3",
    "sphinx>=8.1.3",
    "toml>=0.10.2",
]


[tool.setuptools.packages.find]
include = ["anchor*"]

[tool.setuptools.package-data]
anchor = ["templates/anchor/**/*"]

[tool.setuptools.dynamic]
version = {attr = "anchor.__version__"}
